Replace ImportFromYamlAction with UpgradeAction
Having the update action be called ImportFromYamlAction doesn't make much sense
when it does more than just migrate yaml state. This commit rearanges the
preexisting code to better allow for extension by adding things to the new
UpgradeAction.

UpdateAction currently just makes the same calls the old ImportFromYamlAction
did just in a different place.
