+++ cli/src/action/database/upgrade/mod.rs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//! Provides database upgrade functionality
+
 mod yaml;
 
 use std::path::PathBuf;
@@ -26,7 +28,31 @@ use crate::action::database::{PgConnection, SplinterEnvironment};
 use crate::diesel::Connection;
 use crate::error::CliError;
 
-pub use yaml::ImportFromYamlAction;
+use super::Action;
+
+/// The overarching Action possibly containing multiple upgrade actions
+pub struct UpgradeAction;
+
+impl Action for UpgradeAction {
+    fn run<'a>(&mut self, arg_matches: Option<&ArgMatches<'a>>) -> Result<(), CliError> {
+        let state_dir = get_state_dir(arg_matches)?;
+        let database_uri = get_database_uri(arg_matches)?;
+        info!("Upgrading splinterd state");
+        info!(
+            "Source yaml state directory: {}",
+            state_dir.to_string_lossy()
+        );
+        info!("Destination database uri: {}", database_uri);
+        info!("Loading YAML datastore... ");
+        let store_factory = splinter::store::create_store_factory(database_uri).map_err(|err| {
+            CliError::ActionError(format!("failed to initialized store factory: {}", err))
+        })?;
+
+        let db_store = store_factory.get_admin_service_store();
+        yaml::import_yaml_state_to_database(state_dir, &*db_store)?;
+        Ok(())
+    }
+}
 
 /// Gets the path of splinterd's state directory
 ///
